moved files from dist to src
